MF-954 - Rename pkg/redis to pkg/events
